// src/app/cib-components/cib-table/models/config.model.ts
export class CIBTableConfig {
    columns: Array<CIBTableColumn> = [];
    data: Array<any> = [];
    pageSize?: number;
    totalRecords!: number;
    selection!: boolean;
    clientPagination?: boolean;
    pageSizeOptions?: Array<number> = [];
}

export class CIBTableColumn {
    key!: string;
    displayName!: string;
    type?: ColumnType = ColumnType.label;
    icon?: string;
    sortable?: boolean;
    UUID?: string = '';
    minWidth?: number;
    extraText?: string;
    callBackFn?: (param: any) => boolean | undefined;
}

export class CIBTableQuery {
    sortKey!: string;
    sortDirection!: SortDirection;
    pageIndex: number = 0;
    pageSize: number = 5;
}

export enum SortDirection {
    asc = 'ASC',
    desc = 'DESC',
}

export enum ColumnType {
    link = 'link',
    button = 'button',
    icon = 'icon',
    label = 'label',
    amount = 'amount',
    approve = 'approve',
    reject = 'reject',
    date = 'date',
    status = 'status',
    attachment = 'attachment',
    switch = 'switch',
    toolTip = 'toolTip',
    html = 'html',
}